[modify]:[Base62Util] - int -> long - int -> long - Param Long.MAX_VALUE add

+++ src/main/java/com/example/shorturl/common/util/Base62Util.java
@@ -6,21 +6,21 @@ import lombok.experimental.UtilityClass;
 public class Base62Util {
     private static final char[] BASE62 = "abcdefghijklmnopqrstuvwxyz1234567890ABCDEFGHIJKLMNOPQRSTUVWXYZ".toCharArray();
 
-    public static String encode(int value) {
+    public static String encode(long value) {
         final StringBuilder sb = new StringBuilder();
         do {
-            int i = value % 62;
+            int i = (int) (value % 62);
             sb.append(BASE62[i]);
             value /= 62;
         } while (value > 0);
         return sb.toString();
     }
 
-    public static int decode(String value) {
-        int result=0;
-        int power=1;
+    public static long decode(String value) {
+        long result=0;
+        long power=1;
         for (int i = 0; i < value.length(); i++) {
-            int digit = new String(BASE62).indexOf(value.charAt(i));
+            long digit = new String(BASE62).indexOf(value.charAt(i));
             result += digit * power;
             power *= 62;
         }
